requirements.txt
langchain
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
python-magic-bin
OpenAI
torch==2.3.1
dill 
langchain_groq 
langchain_community 
streamlit 
python-dotenv 
scikit-learn 
numpy 
faiss-cpu